Split code into folders, split up controller
