--- demo/demo/src/main/java/com/example/demo/AccountRepository.java
@@ -1,6 +1,0 @@
-package com.example.demo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface AccountRepository extends JpaRepository<Account, Long> {
-}
