--- demo/demo/src/main/java/com/example/demo/RoomRepository.java
@@ -1,6 +1,0 @@
-package com.example.demo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface RoomRepository extends JpaRepository<Room, Long> {
-}
