+++ demo/demo/src/main/java/com/example/demo/repository/MessageRepository.java
@@ -1,10 +1,11 @@
-package com.example.demo;
+package com.example.demo.repository;
 
+import com.example.demo.model.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-interface MessageRepository extends JpaRepository<Message, Long> {
+public interface MessageRepository extends JpaRepository<Message, Long> {
     List<Message> findByRoomId(Long roomId);
 
     List<Message> findByAccountId(Long accountId);
