+++ demo/demo/src/test/java/com/example/demo/AccountControllerTests.java
@@ -1,5 +1,7 @@
 package com.example.demo;
 
+import com.example.demo.model.Account;
+import com.example.demo.repository.AccountRepository;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,68 +9,27 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 import java.time.LocalDate;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Locale;
+import java.util.Optional;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
-public class AppControllerTests {
+public class AccountControllerTests {
+
     @Autowired
     private MockMvc mockMvc;
 
     @MockBean
-    private RoomRepository roomRepository;
-
-    @MockBean
     private AccountRepository accountRepository;
-
-    @Test
-    public void allRoomsShouldCallRepository() throws Exception {
-        when(roomRepository.findAll()).thenReturn(new ArrayList<>());
-        this.mockMvc.perform(MockMvcRequestBuilders.get("/rooms")).andDo(print()).andExpect(status().isOk()).andExpect(jsonPath("$", Matchers.empty()));
-    }
-
-    @Test
-    public void oneRoomShouldCallRepository() throws Exception {
-        when(roomRepository.findById(1L)).thenReturn(Optional.of(new Room("1", "1")));
-        this.mockMvc.perform(MockMvcRequestBuilders.get("/rooms/1")).andDo(print()).andExpect(status().isOk()).andExpect(jsonPath("$.topic").value("1"));
-    }
-
-    @Test
-    public void oneRoomShouldErrorIfNoMatch() throws Exception {
-        when(roomRepository.findById(1L)).thenReturn(Optional.empty());
-        this.mockMvc.perform(MockMvcRequestBuilders.get("/rooms/1")).andDo(print()).andExpect(status().is4xxClientError());
-    }
-
-    @Test
-    public void newRoomShouldCreateRoom() throws Exception {
-        Room room = new Room("testTopic", "123");
-        when(roomRepository.save(room)).thenReturn(room);
-
-        this.mockMvc.perform(MockMvcRequestBuilders.post("/rooms").contentType("application/json").content("{\"hostId\":\"123\", \"topic\":\"testTopic\"}")).andDo(print()).andExpect(status().isOk()).andExpect(jsonPath("$.hostId").value("123"));
-    }
-
-    @Test
-    public void updateRoomShouldUpdate() throws Exception {
-        when(roomRepository.findById(1L)).thenReturn(Optional.of(new Room("testTopic", "456")));
-        this.mockMvc.perform(MockMvcRequestBuilders.put("/rooms/1").contentType("application/json").content("{\"hostId\":\"123\", \"topic\":\"testTopic\"}")).andDo(print()).andExpect(status().isOk()).andExpect(jsonPath("$.hostId").value("123"));
-    }
-
-    @Test
-    public void updateRoomShouldErrorIfNoMatch() throws Exception {
-        when(roomRepository.findById(1L)).thenReturn(Optional.empty());
-        this.mockMvc.perform(MockMvcRequestBuilders.put("/rooms/1").contentType("application/json").content("{\"hostId\":\"123\", \"topic\":\"testTopic\"}")).andDo(print()).andExpect(status().is4xxClientError());
-    }
 
     @Test
     public void allAccountsShouldCallRepository() throws Exception {
@@ -91,7 +52,7 @@ public class AppControllerTests {
     @Test
     public void newAccountShouldCreateNewAccount() throws Exception {
         this.mockMvc.perform(MockMvcRequestBuilders.post("/accounts").contentType("application/json").content("{\"email\":\"amy@gmail.com\", \"firstName\":\"Amy\", \"lastName\":\"Murphy\", \"username\":\"amurphy1\", \"country\":\"en_IE\", \"dob\":\"1980_1_1\"}"
-                .getBytes()))
+                        .getBytes()))
                 .andDo(print())
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.username").value("amurphy1"));
@@ -101,7 +62,7 @@ public class AppControllerTests {
     public void updateAccountShouldChangeSavedAccount() throws Exception {
         when(accountRepository.findById(2L)).thenReturn(Optional.of(new Account("bob@gmail.com", "Bob", "Dylan", "bdylan1", new Locale("en", "IE"), LocalDate.of(1960, 1, 1))));
         this.mockMvc.perform(MockMvcRequestBuilders.put("/accounts/2").contentType("application/json").content("{\"email\":\"bob@gmail.com\", \"firstName\":\"Bob\", \"lastName\":\"Dylan\", \"username\":\"dylan1bob\", \"country\":\"en_IE\", \"dob\":\"1960_1_1\"}"
-                .getBytes()))
+                        .getBytes()))
                 .andDo(print())
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.username").value("dylan1bob"));
@@ -112,7 +73,7 @@ public class AppControllerTests {
     public void updateAccountShouldErrorIfNoMatch() throws Exception {
         when(accountRepository.findById(2L)).thenReturn(Optional.empty());
         this.mockMvc.perform(MockMvcRequestBuilders.put("/accounts/2").contentType("application/json").content("{\"email\":\"bob@gmail.com\", \"firstName\":\"Bob\", \"lastName\":\"Dylan\", \"username\":\"dylan1bob\", \"country\":\"en_IE\", \"dob\":\"1960_1_1\"}"
-                .getBytes()))
+                        .getBytes()))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
     }
