+++ demo/demo/src/test/java/com/example/demo/ChatroomApplicationTests.java
@@ -1,5 +1,7 @@
 package com.example.demo;
 
+import com.example.demo.controller.AccountController;
+import com.example.demo.controller.RoomController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -10,10 +12,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 @SpringBootTest
 class ChatroomApplicationTests {
 	@Autowired
-	private AppController appController;
+	private AccountController accountController;
+
+	@Autowired
+	private RoomController roomController;
 	@Test
 	void contextLoads() throws Exception {
-		assertThat(appController).isNotNull();
+		assertThat(accountController).isNotNull();
+		assertThat(roomController).isNotNull();
 	}
 
 }
